requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt
drf-yasg==1.21.8
exceptiongroup==1.2.2
flake8==7.1.1
importlib_metadata==8.4.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.1
